5.19

+++ paddle/phi/kernels/cpu/graph_khop_sampler_kernel.cc
@@ -365,7 +365,7 @@ void GraphKhopSamplerKernel(const Context& dev_ctx,
   }
 
   int64_t num_sample_edges = std::accumulate(
-      dst_sample_counts_merge.begin(), dst_sample_counts_merge.end(), 0);
+      dst_sample_counts_merge.begin(), dst_sample_counts_merge.end(), static_cast<int64_t>(0));
   PADDLE_ENFORCE_EQ(
       src_merge.size(),
       num_sample_edges,
